@@ -1,6 +1,6 @@
 {
   "name": "react-addons-linked-state-mixin",
-  "version": "15.5.2",
+  "version": "15.6.0",
   "main": "index.js",
   "repository": "facebook/react",
   "keywords": [
@@ -13,7 +13,11 @@
     "object-assign": "^4.1.0"
   },
   "scripts": {
-    "test": "jest"
+    "test": "TEST_ENTRY=./index.js jest",
+    "build:dev": "NODE_ENV=development webpack && TEST_ENTRY=./react-addons-linked-state-mixin.js jest",
+    "build:prod": "NODE_ENV=production webpack && NODE_ENV=production TEST_ENTRY=./react-addons-linked-state-mixin.min.js jest",
+    "build": "npm run build:dev && npm run build:prod && node ../postbuild.js",
+    "prepublish": "npm test && npm run build"
   },
   "files": [
     "LICENSE",
@@ -24,9 +28,10 @@
     "react-addons-linked-state-mixin.min.js"
   ],
   "devDependencies": {
+    "create-react-class": "^15.5.4",
     "jest": "^19.0.2",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "15.4.2",
-    "react-dom": "^15.4.2"
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "webpack": "^2.6.1"
   }
 }
