@@ -12,48 +12,93 @@
 'use strict';
 
 var React;
-var ReactDOM;
 var createReactFragment;
 
-// For testing DOM Fiber.
-global.requestAnimationFrame = function(callback) {
-  setTimeout(callback);
+// Catch stray warnings
+var env = jasmine.getEnv();
+var callCount = 0;
+var oldError = console.error;
+var newError = function() {
+  callCount++;
+  oldError.apply(this, arguments);
 };
-
-global.requestIdleCallback = function(callback) {
-  setTimeout(() => {
-    callback({ timeRemaining() { return Infinity; } });
+console.error = newError;
+env.beforeEach(() => {
+  callCount = 0;
+  jasmine.addMatchers({
+    toBeReset() {
+      return {
+        compare(actual) {
+          if (actual !== newError && !jasmine.isSpy(actual)) {
+            return {
+              pass: false,
+              message: 'Test did not tear down console.error mock properly.'
+            };
+          }
+          return {pass: true};
+        }
+      };
+    },
+    toNotHaveBeenCalled() {
+      return {
+        compare(actual) {
+          return {
+            pass: callCount === 0,
+            message: 'Expected test not to warn. If the warning is expected, mock ' +
+              "it out using spyOn(console, 'error'); and test that the " +
+              'warning occurs.'
+          };
+        }
+      };
+    }
   });
-};
+});
+env.afterEach(() => {
+  expect(console.error).toBeReset();
+  expect(console.error).toNotHaveBeenCalled();
+});
 
-const expectDev = function expectDev(actual) {
+// Suppress warning expectations for prod builds
+function suppressDevMatcher(obj, name) {
+  const original = obj[name];
+  obj[name] = function devMatcher() {
+    try {
+      original.apply(this, arguments);
+    } catch (e) {
+      // skip
+    }
+  };
+}
+function expectDev(actual) {
   const expectation = expect(actual);
-  if (global.__suppressDevFailures) {
-    Object.keys(expectation).forEach((name) => {
-      wrapDevMatcher(expectation, name);
-      wrapDevMatcher(expectation.not, name);
+  if (process.env.NODE_ENV === 'production') {
+    Object.keys(expectation).forEach(name => {
+      suppressDevMatcher(expectation, name);
+      suppressDevMatcher(expectation.not, name);
     });
   }
   return expectation;
-};
+}
 
 describe('createReactFragment', () => {
   beforeEach(() => {
-    jest.resetModules()
+    jest.resetModules();
 
     React = require('react');
-    ReactDOM = require('react-dom');
-    createReactFragment = require('./index');
+    createReactFragment = require(process.env.TEST_ENTRY);
   });
 
   it('warns for numeric keys on objects as children', () => {
     spyOn(console, 'error');
 
-    createReactFragment({1: React.createElement('span'), 2: React.createElement('span')});
+    createReactFragment({
+      1: React.createElement('span'),
+      2: React.createElement('span')
+    });
 
     expectDev(console.error.calls.count()).toBe(1);
     expectDev(console.error.calls.argsFor(0)[0]).toContain(
-      'Child objects should have non-numeric keys so ordering is preserved.',
+      'Child objects should have non-numeric keys so ordering is preserved.'
     );
   });
 
@@ -62,7 +107,7 @@ describe('createReactFragment', () => {
     createReactFragment(null);
     expectDev(console.error.calls.count()).toBe(1);
     expectDev(console.error.calls.argsFor(0)[0]).toContain(
-      'React.addons.createFragment only accepts a single object.',
+      'React.addons.createFragment only accepts a single object.'
     );
   });
 
@@ -71,7 +116,7 @@ describe('createReactFragment', () => {
     createReactFragment([]);
     expectDev(console.error.calls.count()).toBe(1);
     expectDev(console.error.calls.argsFor(0)[0]).toContain(
-      'React.addons.createFragment only accepts a single object.',
+      'React.addons.createFragment only accepts a single object.'
     );
   });
 
@@ -81,7 +126,7 @@ describe('createReactFragment', () => {
     expectDev(console.error.calls.count()).toBe(1);
     expectDev(console.error.calls.argsFor(0)[0]).toContain(
       'React.addons.createFragment does not accept a ReactElement without a ' +
-        'wrapper object.',
+        'wrapper object.'
     );
   });
 });
