@@ -79,7 +79,7 @@ var ReactStateSetters = {
     // Memoize the setters.
     var cache = component.__keySetters || (component.__keySetters = {});
     return cache[key] || (cache[key] = createStateKeySetter(component, key));
-  },
+  }
 };
 
 function createStateKeySetter(component, key) {
@@ -131,7 +131,7 @@ ReactStateSetters.Mixin = {
    */
   createStateKeySetter: function(key) {
     return ReactStateSetters.createStateKeySetter(this, key);
-  },
+  }
 };
 
 /**
@@ -152,7 +152,7 @@ var LinkedStateMixin = {
       this.state[key],
       ReactStateSetters.createStateKeySetter(this, key)
     );
-  },
+  }
 };
 
 module.exports = LinkedStateMixin;
