@@ -12,36 +12,71 @@
 'use strict';
 
 let LinkedStateMixin;
+let createReactClass;
 let React;
 let ReactDOM;
 let ReactTestUtils;
 
-// For testing DOM Fiber.
-global.requestAnimationFrame = function(callback) {
-  setTimeout(callback);
+// Catch stray warnings
+var env = jasmine.getEnv();
+var callCount = 0;
+var oldError = console.error;
+var newError = function() {
+  callCount++;
+  oldError.apply(this, arguments);
 };
-
-global.requestIdleCallback = function(callback) {
-  setTimeout(() => {
-    callback({ timeRemaining() { return Infinity; } });
+console.error = newError;
+env.beforeEach(() => {
+  callCount = 0;
+  jasmine.addMatchers({
+    toBeReset() {
+      return {
+        compare(actual) {
+          if (actual !== newError && !jasmine.isSpy(actual)) {
+            return {
+              pass: false,
+              message: 'Test did not tear down console.error mock properly.'
+            };
+          }
+          return {pass: true};
+        }
+      };
+    },
+    toNotHaveBeenCalled() {
+      return {
+        compare(actual) {
+          return {
+            pass: callCount === 0,
+            message: 'Expected test not to warn. If the warning is expected, mock ' +
+              "it out using spyOn(console, 'error'); and test that the " +
+              'warning occurs.'
+          };
+        }
+      };
+    }
   });
-};
+});
+env.afterEach(() => {
+  expect(console.error).toBeReset();
+  expect(console.error).toNotHaveBeenCalled();
+});
 
 describe('LinkedStateMixin', () => {
   beforeEach(() => {
-    jest.resetModules()
+    jest.resetModules();
 
+    createReactClass = require('create-react-class');
     React = require('react');
     ReactDOM = require('react-dom');
-    ReactTestUtils = require('react-addons-test-utils');
-    LinkedStateMixin = require('./index');
+    ReactTestUtils = require('react-dom/test-utils');
+    LinkedStateMixin = require(process.env.TEST_ENTRY);
   });
 
   // https://facebook.github.io/react/docs/two-way-binding-helpers.html#linkedstatemixin-before-and-after
   it('should work with valueLink', () => {
     spyOn(console, 'error'); // Ignore deprecated valueLink message for now
 
-    const WithLink = React.createClass({
+    const WithLink = createReactClass({
       mixins: [LinkedStateMixin],
       getInitialState: function() {
         return {message: 'Hello!'};
@@ -66,7 +101,7 @@ describe('LinkedStateMixin', () => {
 
   // https://facebook.github.io/react/docs/two-way-binding-helpers.html#linkedstatemixin-without-valuelink
   it('should work without valueLink', () => {
-    const WithoutLink = React.createClass({
+    const WithoutLink = createReactClass({
       mixins: [LinkedStateMixin],
       getInitialState: function() {
         return {message: 'Hello!'};
@@ -76,7 +111,9 @@ describe('LinkedStateMixin', () => {
         var handleChange = function(e) {
           valueLink.requestChange(e.target.value);
         };
-        return <input type="text" value={valueLink.value} onChange={handleChange} />;
+        return (
+          <input type="text" value={valueLink.value} onChange={handleChange} />
+        );
       }
     });
 
