@@ -9,15 +9,67 @@
 
 'use strict';
 
-var warning = require('fbjs/lib/warning');
+var lowPriorityWarning = function lowPriorityWarning() {};
+
+if (process.env.NODE_ENV !== 'production') {
+  var printWarning = function(format) {
+    for (
+      var _len = arguments.length,
+        args = Array(_len > 1 ? _len - 1 : 0),
+        _key = 1;
+      _key < _len;
+      _key++
+    ) {
+      args[_key - 1] = arguments[_key];
+    }
+
+    var argIndex = 0;
+    var message =
+      'Warning: ' +
+      format.replace(/%s/g, function() {
+        return args[argIndex++];
+      });
+    if (typeof console !== 'undefined') {
+      console.warn(message);
+    }
+    try {
+      // --- Welcome to debugging React ---
+      // This error was thrown as a convenience so that you can use this stack
+      // to find the callsite that caused this warning to fire.
+      throw new Error(message);
+    } catch (x) {}
+  };
+
+  lowPriorityWarning = function(condition, format) {
+    if (format === undefined) {
+      throw new Error(
+        '`warning(condition, format, ...args)` requires a warning ' +
+          'message argument'
+      );
+    }
+    if (!condition) {
+      for (
+        var _len2 = arguments.length,
+          args = Array(_len2 > 2 ? _len2 - 2 : 0),
+          _key2 = 2;
+        _key2 < _len2;
+        _key2++
+      ) {
+        args[_key2 - 2] = arguments[_key2];
+      }
+
+      printWarning.apply(undefined, [format].concat(args));
+    }
+  };
+}
 
 // This package has been deprecated in NPM as of version 15.5.0
 // But NPM deprecation warnings are easy to overlook
 // So a more explicit runtime warning seemed appropriate
-warning(
+lowPriorityWarning(
   false,
   'ReactTestUtils has been moved to react-dom/test-utils. ' +
-  'Update references to remove this warning.'
+    'Update references to remove this warning.'
 );
 
 module.exports = require('react-dom/lib/ReactTestUtils');
