@@ -11,21 +11,54 @@
 
 'use strict';
 
-// Polyfill for testing DOM Fiber.
-global.requestAnimationFrame = function(callback) {
-  setTimeout(callback);
+// Catch stray warnings
+var env = jasmine.getEnv();
+var callCount = 0;
+var oldError = console.error;
+var newError = function() {
+  callCount++;
+  oldError.apply(this, arguments);
 };
-
-global.requestIdleCallback = function(callback) {
-  setTimeout(() => {
-    callback({ timeRemaining() { return Infinity; } });
+console.error = newError;
+env.beforeEach(() => {
+  callCount = 0;
+  jasmine.addMatchers({
+    toBeReset() {
+      return {
+        compare(actual) {
+          if (actual !== newError && !jasmine.isSpy(actual)) {
+            return {
+              pass: false,
+              message: 'Test did not tear down console.error mock properly.'
+            };
+          }
+          return {pass: true};
+        }
+      };
+    },
+    toNotHaveBeenCalled() {
+      return {
+        compare(actual) {
+          return {
+            pass: callCount === 0,
+            message: 'Expected test not to warn. If the warning is expected, mock ' +
+              "it out using spyOn(console, 'error'); and test that the " +
+              'warning occurs.'
+          };
+        }
+      };
+    }
   });
-};
+});
+env.afterEach(() => {
+  expect(console.error).toBeReset();
+  expect(console.error).toNotHaveBeenCalled();
+});
 
 function noop() {}
 
 // Tests adapted from ReactComponentWithPureRendererMixin and ReactPureComponent tests
-describe('LinkedStateMixin', function() {
+describe('LinkedInput', function() {
   let LinkedInput;
   let React;
   let ReactDOM;
@@ -33,7 +66,7 @@ describe('LinkedStateMixin', function() {
   beforeEach(function() {
     React = require('react');
     ReactDOM = require('react-dom');
-    LinkedInput = require('./index');
+    LinkedInput = require(process.env.TEST_ENTRY);
   });
 
   it('should basically work', function() {
@@ -42,7 +75,7 @@ describe('LinkedStateMixin', function() {
     const container = document.createElement('div');
     const component = ReactDOM.render(
       React.createElement(LinkedInput, {
-        value: "foo",
+        value: 'foo',
         onChange: noop
       }),
       container
